chore: add opt-in quality of life development improvements
This is opt-in behavior, meaning it's not enabled by default, as
to not hinder the team and new contributors. However, I strongly
recommend everyone to do opt in.

Here's how:

Create a file called `.opt-in` at the root of the repository,
with the options you wish to enable (one per line).

- `pre-commit`:

  * Source code files are formatted with Prettier automatically.
  * Source code files are checked with TSLint and ESLint automatically.

- `commit-msg`:
  * Commit messages are validated to follow a convention.
  See <https://www.conventionalcommits.org/en/v1.0.0-beta.2/> for
  details.

Please note, if you do opt-in, that since these are new practices, not
enforced before, there's a good chance you might want a workaround if
you're getting warnings from unrelated areas of your current task.
While we encourage you to fix as many as you can, you can always
opt-out in an case by case basis by running `git commit -n`.