refactor: log error when scene collection fails to load
If we fail to load an scene collection, the behavior appears to be
trying to restore from backup. This causes the loading of application
state to be invoked a second time, with unexpected side-effects like
duplicate scenes. This commit addresses the visibility part by
displaying an error message for developers as opposed to swallowing the
error entirely. Prior to this, we addressed the second part of one of
those visible effects by adding uniqueness checks in the scenes API.