refactor(hotkeys): fortify checking, update mappings, details follow
* Fortifies checking for a source's type in case it cannot be found in
the service (it has happened).
* Fix action names for `TOGGLE_MUTE` and `TOGGLE_UNMUTE`.
* Track added hotkeys on a set for the time being, ensuring we don't
add duplicates coming from OBS that we're already handling.
* Add explicit check for whether the returned item belongs to a scene
hotkey.

Please notice that currently identifying which kind of hotkey we got
from the backend is factorial as we don't seem to have a better way to
get this info from the backend. The number of hotkeys affected by this
is minimal (the main reason we're sticking to the former loops for now)
but could still be significant. I have not, however, profiled this.